[dev] Math functions left into another module
