+++ css/chatbot.css
@@ -10,13 +10,32 @@ main.information__main
    min-height : max-content;
 }
 
+.home-router__p
+{
+   display : none;
+}
+
+.chat-top__div
+{
+   width : 100vw;
+   height : 11vh;
+   text-align : right;
+   background-color : var(--white);
+}
+
+.home-router__button,
+.document-panel__button
+{
+   height : 2rem;
+   width : 2rem;
+
+}
 
 .chat-body__div
 {
+   width : 100vw;
    height: 88%;
-   width : 94vw;
    margin : auto;
-   border-radius : 35px;
    background-color : var(--white);
    position : relative;
    overflow-x : hidden;
@@ -24,8 +43,8 @@ main.information__main
 
 .chat-configuration__article
 {
-   height: 50%;
    width : 100%;
+   height: 50%;
    position : absolute;
    background-color : var(--aqua);
    border-radius : 35px;
@@ -45,7 +64,7 @@ main.information__main
 
 .chat-bubble-jar__div
 {
-   height: 89%;
+   height: 90%;
    padding: 1.2rem 2.35rem;
    display: flex;
    flex-direction: column;
@@ -102,13 +121,14 @@ main.information__main
 {
    width : 78%;
    padding : 1.3rem 2rem;
-   border: 1px solid var(--almost-black);
+   margin : auto auto auto -20px;
+   border: none;
    border-radius: 22px;
-   font-size : 2.7rem;
    color : var(--grey);
+   font-size : 2.7rem;
    overflow-y: scroll;
    word-wrap: break-word;
-   line-height : 30px;
+   line-height : 5rem;
 }
 
 #chat-submitter>input[type="submit"]
