+++ src/chatbot.js
@@ -1,10 +1,12 @@
-const chat = document.getElementById("chat-submitter");
-const file = document.getElementById("document-submitter");
+import { cosineSimilarity } from "./math.js";
+
 
 const PDF_EXTRACTOR = "https://servicenuruk.realitynear.org:7725/document";
 const VECTOR_GENERATOR = "https://servicenuruk.realitynear.org:7726/vectorize";
 const PROMPT_END = "https://servicenuruk.realitynear.org:7726/ask";
 
+const chat = document.getElementById("chat-submitter");
+const file = document.getElementById("document-submitter");
 const jar = document.getElementsByClassName("chat-bubble-jar__div")[0];
 
 
@@ -35,24 +37,28 @@ function pushToJar( type, content )
       stickyScroll( jar );
 }
 
-
-chat.addEventListener("submit", async event => {
-
+function firstLine( event )
+{
    event.preventDefault();
 
-   let answer = await ask( event.target.message.value );
-
-   pushToJar( "sender", event.target.message.value );
+   senderProcess(event.target.message.value);
 
    event.target.reset();
+}
+
+
+async function senderProcess( message )
+{
+   pushToJar( "sender", message );
+
+   let answer = await ask( message );
 
    pushToJar( "receiver", answer );
+}
 
-});
 
-
-file.addEventListener("submit", async event => {
-
+async function documentProcessor( event )
+{
    event.preventDefault();
 
    let form = new FormData( event.target );
@@ -71,7 +77,7 @@ file.addEventListener("submit", async event => {
    localStorage.setItem('pages', keys);
 
    let docEmbedding = await vectorizeDocument( keys );
-});
+}
 
 /**
  * Generates and saves all the embeddings per page of the stored document
@@ -103,7 +109,7 @@ function buildContext( sections, vectorX )
       let container = localStorage.getItem(`e.${page}`);
       let vectorY =  JSON.parse(container);
 
-      let similarity = cosineSimilarity( vectorX, vectorY, vectorY.length );
+      let similarity = math.cosineSimilarity( vectorX, vectorY, vectorY.length );
 
       results.push( similarity );
    }
@@ -205,113 +211,3 @@ function retrievePages()
    return localStorage.getItem("pages").split(",");
 }
 
-/**
- * Ranks the first top values of an array of numbers
- */
-function dynamicRank( array, top = 2 )
-{
-   const len = array.length;
-
-   let guns = new Array(top);
-
-   let leftWing = 0, rightWing = leftWing + 1;
-
-   while( rightWing < len)
-   {
-      const underdog = array[leftWing] > array[rightWing] ? leftWing : rightWing;
-
-      for(let i = 0 ; i < top; i++)
-      {
-         if(guns[i] == undefined)
-         {
-            guns[i] = underdog;
-            break;
-         }
-
-         if(array[underdog] > array[guns[i]])
-         {
-            let hold = guns[i];
-            guns[i] = underdog;
-            guns[i+1] = hold;
-            break;
-         }
-      }
-
-      leftWing++;
-      rightWing = leftWing + 1;
-   }
-
-   guns.pop(); // clean last element;
-
-   return guns;
-}
-
-/**
- * Finds the dot product between two vectors of the same dimension
- */
-function dotProduct( x, y, length )
-{
-   let sum = 0;
-
-   for( let i = 0; i < length; i++ )
-   {
-      sum += x[ i ] * y[ i ];
-   }
-
-   return sum;
-}
-
-/**
- * Calculates the Euclidean norm for a vector
- */
-function euclideanNorm( vector )
-{
-   /** Loop configuration **/
-   let i = 0; const length = vector.length;
-
-   /** Equation terms **/
-   let t = 0, ratio = 0, s = 1, unsigned = 0;
-
-   while( i < length)
-   {
-      if(vector[i] == 0)
-      {
-         i++;
-         continue;
-      }
-
-      unsigned = vector[i] < 0 ? -vector[i] : vector[i];
-
-      i++;
-
-      if( unsigned > t )
-      {
-         ratio = t / unsigned;
-         s = 1 + s*ratio**2;
-
-         /** Override t for future **/
-         t = unsigned;
-         continue;
-      }
-
-      ratio = unsigned / t;
-      s = s + ratio**2;
-   }
-
-   return (t * Math.sqrt(s));
-}
-
-/**
- * Takes the cosine similarity of two vector of the same dimensions
- */
-function cosineSimilarity( x, y, dimension )
-{
-   const dot = dotProduct( x, y, dimension );
-
-   const l2normX = euclideanNorm( x );
-   const l2normY = euclideanNorm( y );
-
-   return dot / ( l2normX * l2normY );
-}
-
-export { dynamicRank, euclideanNorm, cosineSimilarity, pushToJar, retrievePages };
